+++ prov/slurm/slurm_node.go
@@ -55,6 +55,12 @@ func generateNodeAllocation(ctx context.Context, kv *api.KV, cfg config.Configur
 		node.memory = re.FindString(strings.Replace(memory.RawString(), " ", "", -1))
 	}
 
+	// Get user credentials from capability endpoint credentials property, if values are provided
+	node.credentials, err = getUserCredentials(kv, deploymentID, nodeName, "endpoint", "credentials")
+	if err != nil {
+		return err
+	}
+
 	// Set the job name property
 	// first: with the prop
 	jobName, err := deployments.GetNodePropertyValue(kv, deploymentID, nodeName, "job_name")
@@ -96,6 +102,10 @@ func generateNodeAllocation(ctx context.Context, kv *api.KV, cfg config.Configur
 	if partition != nil {
 		node.partition = partition.RawString()
 	}
+	// ToDo - #281
+	// Check that is userName provided, then password or privateKey provided also
+	// Otherwise, raise error, event ...
+
 	infra.nodes = append(infra.nodes, node)
 	return nil
 }